🚨 Fix compiler warnings

+++ Source/player.cpp
@@ -3131,7 +3131,7 @@ void ApplyPlrDamage(int pnum, int dam, int minHP /*= 0*/, int frac /*= 0*/, int 
 
 	int totalDamage = (dam << 6) + frac;
 	if (totalDamage > 0 && player.pManaShield) {
-		int manaShieldLevel = player._pSplLvl[SPL_MANASHIELD];
+		int8_t manaShieldLevel = player._pSplLvl[SPL_MANASHIELD];
 		if (manaShieldLevel > 0) {
 			totalDamage += totalDamage / -3;
 		}
